+++ learn2learn/vision/datasets/__init__.py
@@ -15,3 +15,5 @@ from .fc100 import FC100
 from .vgg_flowers import VGGFlower102
 from .fgvc_aircraft import FGVCAircraft
 from .cu_birds200 import CUBirds200
+from .describable_textures import DescribableTextures
+from .quickdraw import Quickdraw