+++ CHANGELOG.md
@@ -10,7 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* `l2l.vision.datasets.DescribableTextures`
+* `l2l.vision.datasets.Quickdraw`
+
 ### Changed
+
+* Updated reference for citations.
 
 ### Fixed